move single-use package
